+++ frontend/src/src/App.jsx
@@ -1,0 +1,23 @@
+import React from "react";
+import "@/App.css";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { Toaster } from "@/components/ui/sonner";
+import HomePage from "@/pages/HomePage";
+import PastePage from "@/pages/PastePage";
+import { ThemeProvider } from "@/hooks/useTheme";
+
+function App() {
+  return (
+    <ThemeProvider>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<HomePage />} />
+          <Route path="/:slug" element={<PastePage />} />
+        </Routes>
+      </BrowserRouter>
+      <Toaster position="bottom-right" />
+    </ThemeProvider>
+  );
+}
+
+export default App;